requirements.txt
# FastAPI backend
fastapi==0.111.0
uvicorn==0.30.1

# Deep Learning
tensorflow==2.14.0

# ML Preprocessing
scikit-learn==1.4.1.post1
joblib==1.4.2
numpy==1.26.4
pandas==2.2.2

# Database
SQLAlchemy==2.0.30
psycopg2-binary==2.9.9

# GUI - PyQt5 chatbot
PyQt5==5.15.9
